proxy/requirements.txt
# S3 to Azure Storage Proxy Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
azure-storage-blob==12.19.0
azure-storage-file-share==12.15.0
